+++ src/main/java/act/event/bytecode/SimpleEventListenerByteCodeScanner.java
@@ -91,15 +91,11 @@ public class SimpleEventListenerByteCodeScanner extends AppByteCodeScannerBase {
         @Override
         public MethodVisitor visitMethod(int access, String name, String desc, String signature, String[] exceptions) {
             MethodVisitor mv = super.visitMethod(access, name, desc, signature, exceptions);
-            Type returnType = Type.getReturnType(desc);
-            final boolean isVoid = "V".equals(returnType.toString());
             final boolean isPublicNotAbstract = AsmTypes.isPublicNotAbstract(access);
             Type[] arguments = Type.getArgumentTypes(desc);
             final List<String> paramTypes = new ArrayList<>();
-            if (null != arguments) {
-                for (Type type : arguments) {
-                    paramTypes.add(type.getClassName());
-                }
+            for (Type type : arguments) {
+                paramTypes.add(type.getClassName());
             }
             final String methodName = name;
             final boolean isStatic = AsmTypes.isStatic(access);
@@ -201,9 +197,7 @@ public class SimpleEventListenerByteCodeScanner extends AppByteCodeScannerBase {
                             }
                         };
                     } else if (Async.class.getName().equals(className)) {
-                        if (!isVoid) {
-                            logger.warn("Error found in method %s.%s: @Async annotation cannot be used with method that has return type", className, methodName);
-                        } else if (!isPublicNotAbstract) {
+                        if (!isPublicNotAbstract) {
                             logger.warn("Error found in method %s.%s: @Async annotation cannot be used with method that are not public or abstract method", className, methodName);
                         } else {
                             asyncMethodName = Async.MethodNameTransformer.transform(methodName);